nsqreader/requirements.txt
sqlalchemy>=1.3.20
marshmallow==3.5.1
pynsq==0.8.3  